requirements.txt
bcrypt==4.0.1
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.3
python-dotenv==0.21.0
Werkzeug==2.2.2
zipp==3.11.0
